+++ dartdoc_test/bin/dartdoc_test.dart
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:dartdoc_test/src/command.dart';
+import 'package:dartdoc_test/src/command/command_runner.dart';
 
 Future<void> main(List<String> arguments) async {
   await DartdocTestCommandRunner().run(arguments);
